move deviceMemory / hardwareConcurrency to a library
